+++ tests/signup.test.js
@@ -22,46 +22,6 @@ afterEach(async () => {
   await connection.query('DELETE FROM countries');
 });
 
-describe('POST /signin', () => {
-  beforeEach(async () => {
-    await insertUser(
-      mockedUser.name,
-      mockedUser.email,
-      mockedUser.hashedPassword(),
-      mockedUser.selectedCountry,
-      mockedUser.birthDate
-    );
-    await addNewSession();
-  });
-
-  afterEach(async () => {
-    await deleteUser();
-    await connection.query('DELETE FROM countries');
-  });
-
-  test('Should return status code 200 and a body with user_id and token, if user is registered', async () => {
-    const connectionData = {
-      email: mockedUser.email,
-      password: mockedUser.password,
-    };
-
-    const result = await agent.post('/signin').send(connectionData);
-    expect(result.status).toEqual(200);
-    expect(result.body).toHaveProperty('token');
-  });
-
-  test('Sould return status code 400, if user is not registered or has passed invalid credentials', async () => {
-    const connectionData = {
-      email: mockedUser.email,
-      password: mockedUser.fakePassword(),
-    };
-
-    const result = await agent.post('/signin').send(connectionData);
-    expect(result.status).toEqual(400);
-    expect(result.body).toEqual({ message: 'Usuário ou senha incorretos.' });
-  });
-});
-
 describe('POST /signup', () => {
   beforeEach(async () => {
     await deleteUser();
